initial 'test-installed-packages' support
This should allow us to break out the installation and testing phases of our
packaging process.  We not only comply with Debian packaging standards by
doing this (allowing for the automagic override of the testing phase, for
instance), but also speed things up significantly by enabling our dist tests
to run in parallel. (!)